[Imp] DMF Loader: Offset without note should now work. [Ref] Got rid of some compiler warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@1992 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_gdm.cpp
@@ -169,7 +169,7 @@ bool CSoundFile::ReadGDM(FileReader &file, ModLoadingFlags loadFlags)
 		ChnSettings[i].Reset();
 		if(fileHeader.panMap[i] < 16)
 		{
-			ChnSettings[i].nPan = std::min((fileHeader.panMap[i] * 16) + 8, 256);
+			ChnSettings[i].nPan = static_cast<uint16>(std::min((fileHeader.panMap[i] * 16) + 8, 256));
 		} else if(fileHeader.panMap[i] == 16)
 		{
 			ChnSettings[i].nPan = 128;
